==> 0.StartHere.R <==
###########################################################################
#IMPT: download the train/test data from kaggle and place in /data folder:
#http://www.kaggle.com/c/see-click-predict-fix/data
###########################################################################

#install the relevant packages
install.packages("stringr")
install.packages("caret")
install.packages("data.table")
install.packages("maptools")
install.packages("rgeos")
install.packages("rgdal")
install.packages("sp")
install.packages("leafletR")
install.packages("classInt")
install.packages("spdep")
install.packages("data.table")
